+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## v0.2.0
+- Fix the _populate_with_tags function
 - Add `Tag.Focused` and `Tag.Hovered` messages that fire, if a `Tag` is hovered or focused
 
 ## v0.1.2